+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 #########
 
+3.0.0
+*****
+
+- changed class members from protected to private
+- cs fixes, added codestyle checks
+
+
 2.0.0
 *****